Release Telux_master_v1.48.0
update rits to adopt Misbehavior Detection API changes
enable chrony-sock for cv2x only
Updated telsdk console app data call notification parameters
Updates to C-V2X radio call flows and doxy comments of some APIs.
ISensor renamed to ISensorClient retaining backward compatibility
Add proxy in cv2x_tcp_test tool
Fix Y2038 issue found in QITS
Add dgnss manager callflow
sensor_test_app-Improve sensor feature test menu
audio_console_app enhancements
Added support for displaying SDK release name in test app
Miscellaneous changes in eCall reference application
Add network time information API.
Get telux versions test app
capture doc version in revision history table
Adding Public interfaces for RF Band Info
Added telux_version_app
Added fix for Y2038 issues

CRs-Fixed: 3115663, 3120703

+++ apps/reference/rits/src/qApplication/SafetyApps/safetyapp_util.cpp
@@ -29,7 +29,7 @@
 /*
  *  Changes from Qualcomm Innovation Center are provided under the following license:
  *
- *  Copyright (c) 2021 Qualcomm Innovation Center, Inc. All rights reserved.
+ *  Copyright (c) 2021-2022 Qualcomm Innovation Center, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted (subject to the limitations in the
@@ -76,7 +76,7 @@
 #include <sys/time.h>
 #include <stdbool.h>
 #include "safetyapp_util.h"
- 
+
 /** Maximum number of vehicles including the host that the LDM can support */
 double MAX_MAP_SIZE;
 
@@ -412,7 +412,7 @@ lane_types classify_lane(msg_contents *host, msg_contents *remote)
 }
 
 /*
- * Differs from basic verison. Also checks if the host would approximately travel in the 
+ * Differs from basic verison. Also checks if the host would approximately travel in the
  * same dir of RV
  */
 double time_to_crash_adv(msg_contents *host, msg_contents *remote)
@@ -494,7 +494,7 @@ static __inline uint64_t ntimestamp_now(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000000LL + tv.tv_usec;
+    return (uint64_t)tv.tv_sec * 1000000LL + (uint64_t)tv.tv_usec;
 }
 
 /* From the received secmark, calculate the timestamp*/
